requirements.txt
fastapi
uvicorn
openai
spotipy
python-dotenv
pandas